remove deployments.json from CLI
This will replace the functions related to `deployments.json`
with a function that checks if a specified path is a deployment, that is:
- has a 'config.toml' file
- has a 'channels' directory
- the database specified in the config file exists.

The database existence check is performed via sqlalchemy_utils.

`quetz create` has a new `delete` option instead of the delete confirmation,
and only works if either `copy_conf` or `create_conf` is specified.
`skip_if_exists` is renamed to `exists_ok`, also.

`config_file_name` is removed from both `run` and `create` commands,
`config.toml` is the only allowed config file name.
file specified in `copy_conf` will be renamed to `config.toml` in the deployment directory.

Pathlib objects are used extensively.
we also use a context manager `working_directory`
to ensure we return to the original folder after the run.