fix: check if a channel exists before creating it
`mkdir` in gscfs raises an HTTPError: "You already own this bucket.
Please select another name., 409" if the bucket already exists. I
modified the behavior of `create_channel` to check if the bucket exists
before creating it. This will usually result in the same number of API
calls since most of the time the bucket will exist.