feat: migrate to Pydantic v2 (#656)
* Upgrade to pydantic v2
* Replace deprecated dict with model_dump
* Replace update_forward_refs with model_rebuild
* Fix error message in test
* Fix Packages and ChannelSearch rest models

- platforms is nullable: it should be optional
- Channel description is optional - so should be ChannelSearch description

* Fix Job rest model

item_spec can't be set to required in JobBase and optional in Job
(mypy complains).
Create a new class JobCreate where item_spec is required.
In job, item_spec is optional.

* Add UserOptionalRole

User role is nullable and should be optional in get (but not in set)

* Remove nullable=True in pydantic Field

This isn't used.
To mark a field as nullable, it should be set as Optional or "| None".

Fix deprecation warning: Extra keyword arguments on `Field` is deprecated and will be removed.

* Fix Channel model

keyword should be examples (not example)

* Replace deprecated from_orm with model_validate

* Replace deprecated parse_obj with model_validate

* Fix UserWarning

UserWarning: `pydantic.error_wrappers:ValidationError` has been moved to `pydantic:ValidationError`.

* Fix test_post_new_job_invalid_items_spec

* Fix starlette Deprecation Warning

DeprecationWarning: The `allow_redirects` argument is deprecated. Use `follow_redirects` instead.

* Increase wait time for running task tests

Test failing with postgres. Task still pending after 2.5 seconds.